Extract whiteboard, IEP, and assessment into ES modules
Phase 2 of script.js modularization: extract three major subsystems
into dedicated ES module files, reducing script.js from 4,208 to
2,626 lines (-37.6%).

- modules/whiteboard.js (887 lines): Shelved whiteboard system with
  canvas drawing, toolbar, templates, and coordinate grids. Exports
  initWhiteboardSystem() factory for when feature is un-shelved.

- modules/iep.js (373 lines): IEP accommodation system including
  page-level accommodations, per-response features, goal progress
  notifications, chunked assignment check-ins, break overlay with
  breathing/stretch/grounding exercises, and multiplication chart.
  Uses createIepSystem() factory with dependency injection.

- modules/assessment.js (309 lines): In-chat CAT placement
  assessment flow — pitch, problem loading, answer submission,
  and completion handling. Uses createAssessmentSystem() factory.

All modules use a factory pattern that accepts dependencies from the
DOMContentLoaded closure, keeping the wiring clean and testable.
