+++ public/js/script.js
@@ -8,13 +8,14 @@ import { sessionTracker, initSessionTracking, getActiveSeconds, sendTimeHeartbea
 import { showLevelUpCelebration, triggerXpAnimation as _triggerXpAnimation, updateGamificationDisplay as _updateGamificationDisplay, fetchAndDisplayLeaderboard, loadQuestsAndChallenges, showTutorUnlockCelebration } from './modules/gamification.js';
 import { checkBillingStatus, updateFreeTimeIndicator, showUpgradePrompt, initiateUpgrade } from './modules/billing.js';
 import { audioState, audioQueue, playAudio, processAudioQueue, pauseAudio, resumeAudio, restartAudio, stopAudio, changePlaybackSpeed, resetAudioState, updateAudioControls } from './modules/audio.js';
+import { createIepSystem } from './modules/iep.js';
+import { createAssessmentSystem } from './modules/assessment.js';
+// Whiteboard is shelved for beta — see modules/whiteboard.js to re-enable
 
 // --- Global Variables ---
 let currentUser = null;
 let isPlaying = false;
 let currentAudioSource = null;
-let fabricCanvas = null;
-let whiteboard = null; // New whiteboard instance
 let attachedFile = null;
 let isRapportBuilding = false; // Track if user is in rapport building phase
 
@@ -23,17 +24,6 @@ let messageQueue = {
     queue: [],
     isProcessing: false,
     currentMessageId: 0
-};
-
-// Whiteboard state
-let whiteboardState = {
-    currentTool: 'pen',
-    currentColor: '#000000',
-    brushSize: 3,
-    isDrawing: false,
-    startX: 0,
-    startY: 0,
-    currentShape: null
 };
 
 // Wrappers that pass currentUser to module functions
@@ -144,364 +134,13 @@ document.addEventListener("DOMContentLoaded", () => {
         }
     }
 
-    // --- IEP Accommodation System ---
-    // Applies page-level accommodations at init (from user profile)
-    function applyIepAccommodations(user) {
-        const accom = user?.iepPlan?.accommodations;
-        if (!accom) return;
-
-        const hasAny = Object.values(accom).some(v => v === true || (Array.isArray(v) && v.length > 0));
-        if (!hasAny) return;
-
-        console.log('[IEP] Applying accommodations:', accom);
-
-        // Reduced Distraction: hide leaderboard, quests, streaks, badges sidebar
-        if (accom.reducedDistraction) {
-            document.body.classList.add('iep-reduced-distraction');
-            console.log('[IEP] Reduced distraction mode enabled');
-        }
-
-        // Large Print / High Contrast: force high-contrast theme
-        // IEP accommodation overrides dark mode — high contrast requires light bg + dark text
-        if (accom.largePrintHighContrast) {
-            document.body.classList.remove('dark-mode');
-            document.body.classList.add('iep-high-contrast');
-            console.log('[IEP] High contrast mode enabled (dark mode overridden)');
-        }
-
-        // Calculator Always Available: auto-show calculator button prominently
-        if (accom.calculatorAllowed) {
-            const calcBtn = document.getElementById('sidebar-calculator-btn');
-            if (calcBtn) calcBtn.classList.add('iep-always-visible');
-        }
-
-        // Breaks As Needed: inject persistent break button
-        if (accom.breaksAsNeeded) {
-            injectBreakButton();
-        }
-
-        // Digital Multiplication Chart: inject persistent chart toggle
-        if (accom.digitalMultiplicationChart) {
-            injectMultiplicationChart();
-        }
-
-        // Math Anxiety Support: add body class for gentler animations
-        if (accom.mathAnxietySupport) {
-            document.body.classList.add('iep-anxiety-support');
-        }
-    }
-
-    // Handles per-response IEP features (from chat API response)
-    function handleIepResponseFeatures(iepFeatures) {
-        if (!iepFeatures) return;
-
-        // Auto Read-Aloud: trigger TTS automatically on new AI messages
-        // Uses the same TTS path as hands-free mode (playAudio + generateSpeakableText)
-        if (iepFeatures.autoReadAloud) {
-            const messages = document.querySelectorAll('.message.ai');
-            const latest = messages[messages.length - 1];
-            if (latest && typeof playAudio === 'function' && typeof generateSpeakableText === 'function' && window.TUTOR_CONFIG) {
-                const tutor = window.TUTOR_CONFIG[currentUser?.selectedTutorId] || window.TUTOR_CONFIG['default'];
-                const playBtn = latest.querySelector('.play-audio-btn');
-                if (playBtn) {
-                    playBtn.disabled = true;
-                    playBtn.classList.add('is-loading');
-                }
-                const rawText = latest.querySelector('.message-content')?.textContent || latest.textContent || '';
-                const speakableText = generateSpeakableText(rawText);
-                if (speakableText) {
-                    setTimeout(() => playAudio(speakableText, tutor?.voiceId, latest.id), 300);
-                }
-            } else {
-                console.warn('[IEP] Audio read-aloud accommodation is active but TTS is unavailable');
-            }
-        }
-
-        // Chunked Assignments: show check-in after every chunk of problems
-        if (iepFeatures.chunkedAssignments) {
-            handleChunkedAssignmentCheckIn();
-        }
-    }
-
-    // Show IEP goal progress updates to the student
-    function handleIepGoalUpdates(goalUpdates) {
-        if (!goalUpdates || goalUpdates.length === 0) return;
-
-        goalUpdates.forEach((update, i) => {
-            setTimeout(() => {
-                showIepGoalNotification(update);
-            }, i * 1200); // Stagger if multiple goals updated
-        });
-    }
-
-    function showIepGoalNotification(update) {
-        const notification = document.createElement('div');
-        notification.className = 'iep-goal-notification';
-
-        if (update.completed) {
-            notification.innerHTML = `
-                <div class="iep-goal-notif-icon completed"><i class="fas fa-star"></i></div>
-                <div class="iep-goal-notif-text">
-                    <strong>Goal Complete!</strong>
-                    <span>${escapeHtml(update.description)}</span>
-                </div>
-            `;
-            notification.classList.add('completed');
-        } else {
-            notification.innerHTML = `
-                <div class="iep-goal-notif-icon"><i class="fas fa-chart-line"></i></div>
-                <div class="iep-goal-notif-text">
-                    <strong>+${update.change}% Goal Progress</strong>
-                    <span>${escapeHtml(update.description)}</span>
-                    <div class="iep-goal-notif-bar">
-                        <div class="iep-goal-notif-fill" style="width:${update.newProgress}%"></div>
-                    </div>
-                </div>
-            `;
-        }
-
-        document.body.appendChild(notification);
-
-        // Animate in
-        requestAnimationFrame(() => notification.classList.add('visible'));
-
-        // Auto-dismiss
-        setTimeout(() => {
-            notification.classList.remove('visible');
-            setTimeout(() => notification.remove(), 400);
-        }, 4000);
-
-        // Click to dismiss
-        notification.addEventListener('click', () => {
-            notification.classList.remove('visible');
-            setTimeout(() => notification.remove(), 400);
-        });
-    }
-
-    // --- Chunked Assignments Enforcement ---
-    // Tracks problems attempted in the current chunk and shows a check-in
-    // overlay after every CHUNK_SIZE problems, giving students a structured
-    // pause point as required by the chunkedAssignments accommodation.
-    let iepChunkProblemCount = 0;
-    const IEP_CHUNK_SIZE = 4; // check in after every 4 problems (within the 3-5 range)
-
-    function handleChunkedAssignmentCheckIn() {
-        // Only count when a problem was just answered (problemResult present)
-        // We detect this by checking if the latest AI response follows a user answer
-        const messages = document.querySelectorAll('.message');
-        if (messages.length < 2) return;
-
-        // Look for the pattern: user message then AI message (problem answered)
-        const lastTwo = Array.from(messages).slice(-2);
-        if (lastTwo[0]?.classList.contains('user') && lastTwo[1]?.classList.contains('ai')) {
-            iepChunkProblemCount++;
-        }
-
-        if (iepChunkProblemCount >= IEP_CHUNK_SIZE) {
-            iepChunkProblemCount = 0;
-            showChunkedCheckIn();
-        }
-    }
-
-    function showChunkedCheckIn() {
-        if (document.getElementById('iep-chunk-checkin')) return;
-
-        const overlay = document.createElement('div');
-        overlay.id = 'iep-chunk-checkin';
-        overlay.className = 'iep-break-overlay';
-        overlay.innerHTML = `
-            <div class="iep-break-card">
-                <h2>Nice work on that set!</h2>
-                <p>You just finished ${IEP_CHUNK_SIZE} problems. How are you feeling?</p>
-                <div class="iep-break-activities">
-                    <button class="iep-break-activity" data-choice="continue">
-                        <i class="fas fa-arrow-right"></i>
-                        <span>Keep Going</span>
-                        <small>I'm ready!</small>
-                    </button>
-                    <button class="iep-break-activity" data-choice="break">
-                        <i class="fas fa-pause-circle"></i>
-                        <span>Take a Break</span>
-                        <small>I need a moment</small>
-                    </button>
-                </div>
-            </div>
-        `;
-        document.body.appendChild(overlay);
-
-        overlay.querySelectorAll('.iep-break-activity').forEach(btn => {
-            btn.addEventListener('click', () => {
-                overlay.remove();
-                if (btn.dataset.choice === 'break') {
-                    showBreakOverlay();
-                }
-            });
-        });
-    }
-
-    function escapeHtml(text) {
-        const div = document.createElement('div');
-        div.textContent = text;
-        return div.innerHTML;
-    }
-
-    // Inject a persistent "Take a Break" button for breaksAsNeeded accommodation
-    function injectBreakButton() {
-        if (document.getElementById('iep-break-btn')) return;
-
-        const btn = document.createElement('button');
-        btn.id = 'iep-break-btn';
-        btn.className = 'iep-break-btn';
-        btn.innerHTML = '<i class="fas fa-pause-circle"></i> Take a Break';
-        btn.title = 'Take a brain break';
-        btn.setAttribute('aria-label', 'Take a brain break');
-        btn.addEventListener('click', showBreakOverlay);
-        document.body.appendChild(btn);
-    }
-
-    function showBreakOverlay() {
-        if (document.getElementById('iep-break-overlay')) return;
-
-        const overlay = document.createElement('div');
-        overlay.id = 'iep-break-overlay';
-        overlay.className = 'iep-break-overlay';
-        overlay.innerHTML = `
-            <div class="iep-break-card">
-                <h2>Brain Break Time</h2>
-                <p>Take a moment to relax. Try one of these:</p>
-                <div class="iep-break-activities">
-                    <button class="iep-break-activity" data-activity="breathe">
-                        <i class="fas fa-wind"></i>
-                        <span>Breathe</span>
-                        <small>4-7-8 breathing</small>
-                    </button>
-                    <button class="iep-break-activity" data-activity="stretch">
-                        <i class="fas fa-child"></i>
-                        <span>Stretch</span>
-                        <small>Quick body stretch</small>
-                    </button>
-                    <button class="iep-break-activity" data-activity="countdown">
-                        <i class="fas fa-eye"></i>
-                        <span>5-4-3-2-1</span>
-                        <small>Grounding exercise</small>
-                    </button>
-                </div>
-                <div id="iep-break-exercise" class="iep-break-exercise" style="display:none;"></div>
-                <button class="iep-break-done-btn" id="iep-break-done">
-                    <i class="fas fa-check-circle"></i> I'm Ready to Continue
-                </button>
-            </div>
-        `;
-        document.body.appendChild(overlay);
-
-        // Activity handlers
-        overlay.querySelectorAll('.iep-break-activity').forEach(actBtn => {
-            actBtn.addEventListener('click', () => {
-                const activity = actBtn.dataset.activity;
-                const exerciseArea = document.getElementById('iep-break-exercise');
-                exerciseArea.style.display = 'block';
-
-                if (activity === 'breathe') {
-                    exerciseArea.innerHTML = `
-                        <div class="iep-breathing-circle">
-                            <div class="iep-breathing-dot"></div>
-                        </div>
-                        <p class="iep-breathing-instruction">Breathe in... 4 seconds</p>
-                    `;
-                    runBreathingExercise(exerciseArea);
-                } else if (activity === 'stretch') {
-                    exerciseArea.innerHTML = `
-                        <div class="iep-stretch-steps">
-                            <p>1. Reach your arms up high and stretch</p>
-                            <p>2. Roll your shoulders backward 5 times</p>
-                            <p>3. Turn your head slowly left, then right</p>
-                            <p>4. Shake out your hands</p>
-                        </div>
-                    `;
-                } else if (activity === 'countdown') {
-                    exerciseArea.innerHTML = `
-                        <div class="iep-grounding">
-                            <p><strong>5</strong> things you can <em>see</em></p>
-                            <p><strong>4</strong> things you can <em>touch</em></p>
-                            <p><strong>3</strong> things you can <em>hear</em></p>
-                            <p><strong>2</strong> things you can <em>smell</em></p>
-                            <p><strong>1</strong> thing you can <em>taste</em></p>
-                        </div>
-                    `;
-                }
-            });
-        });
-
-        document.getElementById('iep-break-done').addEventListener('click', () => {
-            overlay.remove();
-        });
-    }
-
-    function runBreathingExercise(container) {
-        const instruction = container.querySelector('.iep-breathing-instruction');
-        const circle = container.querySelector('.iep-breathing-circle');
-        if (!instruction || !circle) return;
-
-        const phases = [
-            { text: 'Breathe in... 4 seconds', class: 'inhale', duration: 4000 },
-            { text: 'Hold... 7 seconds', class: 'hold', duration: 7000 },
-            { text: 'Breathe out... 8 seconds', class: 'exhale', duration: 8000 }
-        ];
-        let phaseIndex = 0;
-        let cycles = 0;
-
-        function nextPhase() {
-            if (cycles >= 3 || !document.getElementById('iep-break-overlay')) return;
-            const phase = phases[phaseIndex];
-            instruction.textContent = phase.text;
-            circle.className = 'iep-breathing-circle ' + phase.class;
-            phaseIndex = (phaseIndex + 1) % phases.length;
-            if (phaseIndex === 0) cycles++;
-            setTimeout(nextPhase, phase.duration);
-        }
-        nextPhase();
-    }
-
-    // Inject a multiplication chart toggle for digitalMultiplicationChart accommodation
-    function injectMultiplicationChart() {
-        if (document.getElementById('iep-mult-chart-btn')) return;
-
-        const btn = document.createElement('button');
-        btn.id = 'iep-mult-chart-btn';
-        btn.className = 'iep-mult-chart-btn';
-        btn.innerHTML = '<i class="fas fa-th"></i> &times; Chart';
-        btn.title = 'Open multiplication chart';
-        btn.setAttribute('aria-label', 'Open multiplication chart');
-        btn.addEventListener('click', toggleMultiplicationChart);
-        document.body.appendChild(btn);
-    }
-
-    function toggleMultiplicationChart() {
-        let chart = document.getElementById('iep-mult-chart');
-        if (chart) {
-            chart.remove();
-            return;
-        }
-
-        chart = document.createElement('div');
-        chart.id = 'iep-mult-chart';
-        chart.className = 'iep-mult-chart';
-
-        let html = '<div class="iep-mult-chart-header"><h3>Multiplication Chart</h3><button class="iep-mult-chart-close" aria-label="Close chart">&times;</button></div>';
-        html += '<table class="iep-mult-table"><thead><tr><th>&times;</th>';
-        for (let i = 1; i <= 12; i++) html += `<th>${i}</th>`;
-        html += '</tr></thead><tbody>';
-        for (let r = 1; r <= 12; r++) {
-            html += `<tr><th>${r}</th>`;
-            for (let c = 1; c <= 12; c++) html += `<td>${r * c}</td>`;
-            html += '</tr>';
-        }
-        html += '</tbody></table>';
-        chart.innerHTML = html;
-        document.body.appendChild(chart);
-
-        chart.querySelector('.iep-mult-chart-close').addEventListener('click', () => chart.remove());
-    }
+    // --- IEP Accommodation System (see modules/iep.js) ---
+    const iep = createIepSystem({
+        playAudio,
+        generateSpeakableText,
+        getCurrentUser: () => currentUser,
+    });
+    const { applyIepAccommodations, handleIepResponseFeatures, handleIepGoalUpdates } = iep;
 
     // --- Core Functions ---
     async function initializeApp() {
@@ -616,920 +255,8 @@ document.addEventListener("DOMContentLoaded", () => {
         }
     }
     
-    function initializeWhiteboard() {
-        if (document.getElementById('tutor-canvas') && window.MathmatixWhiteboard) {
-            // Initialize the new whiteboard system
-            whiteboard = new MathmatixWhiteboard('tutor-canvas', 'whiteboard-panel');
-            window.whiteboard = whiteboard; // Make available globally for layout manager
-            fabricCanvas = whiteboard.canvas; // Keep for backward compatibility
+    // WHITEBOARD SHELVED FOR BETA — see modules/whiteboard.js
 
-            // Setup toolbar button event listeners
-            setupWhiteboardToolbar();
-
-            // Setup toggle button
-            const toggleBtn = document.getElementById('toggle-whiteboard-btn');
-            const whiteboardPanel = document.getElementById('whiteboard-panel');
-            const openWhiteboardBtn = document.getElementById('open-whiteboard-btn');
-
-            if (toggleBtn && whiteboardPanel) {
-                toggleBtn.addEventListener('click', () => {
-                    const isHidden = whiteboardPanel.classList.contains('is-hidden');
-
-                    if (isHidden) {
-                        // Open whiteboard using proper method to trigger layout manager
-                        if (window.whiteboard && typeof window.whiteboard.show === 'function') {
-                            window.whiteboard.show();
-                        } else {
-                            // Fallback if whiteboard not fully initialized
-                            whiteboardPanel.classList.remove('is-hidden');
-                        }
-                        if (openWhiteboardBtn) {
-                            openWhiteboardBtn.classList.add('hidden');
-                        }
-                    } else {
-                        // Close whiteboard using proper method to trigger layout manager
-                        if (window.whiteboard && typeof window.whiteboard.hide === 'function') {
-                            window.whiteboard.hide();
-                        } else {
-                            // Fallback if whiteboard not fully initialized
-                            whiteboardPanel.classList.add('is-hidden');
-                        }
-                        if (openWhiteboardBtn) {
-                            openWhiteboardBtn.classList.remove('hidden');
-                        }
-                    }
-                });
-            }
-
-            if (openWhiteboardBtn && whiteboardPanel) {
-                openWhiteboardBtn.addEventListener('click', () => {
-                    // Open whiteboard using proper method to trigger layout manager
-                    if (window.whiteboard && typeof window.whiteboard.show === 'function') {
-                        window.whiteboard.show();
-                    } else {
-                        // Fallback if whiteboard not fully initialized
-                        whiteboardPanel.classList.remove('is-hidden');
-                    }
-                    openWhiteboardBtn.classList.add('hidden');
-                });
-            }
-
-            console.log('✅ Modern whiteboard initialized');
-        }
-    }
-
-    function setupWhiteboardToolbar() {
-        // Tool buttons
-        const tools = ['select', 'pen', 'highlighter', 'eraser', 'line', 'arrow',
-                      'rectangle', 'circle', 'triangle', 'text'];
-
-        tools.forEach(tool => {
-            const btn = document.getElementById(`tool-${tool}`);
-            if (btn) {
-                btn.addEventListener('click', () => {
-                    // Remove active class from all tool buttons
-                    document.querySelectorAll('.toolbar-btn').forEach(b => b.classList.remove('active'));
-                    btn.classList.add('active');
-                    whiteboard.setTool(tool);
-                });
-            }
-        });
-
-        // Arrow mode dropdown
-        const arrowBtn = document.getElementById('tool-arrow');
-        const arrowModeMenu = document.getElementById('arrow-mode-menu');
-
-        if (arrowBtn && arrowModeMenu) {
-            // Right-click to open mode menu
-            arrowBtn.addEventListener('contextmenu', (e) => {
-                e.preventDefault();
-                e.stopPropagation();
-                arrowModeMenu.style.display = arrowModeMenu.style.display === 'none' ? 'block' : 'none';
-            });
-
-            // Close dropdown when clicking outside
-            document.addEventListener('click', (e) => {
-                if (!arrowBtn.contains(e.target) && !arrowModeMenu.contains(e.target)) {
-                    arrowModeMenu.style.display = 'none';
-                }
-            });
-
-            // Handle arrow mode selection
-            const arrowModeOptions = arrowModeMenu.querySelectorAll('.arrow-mode-option');
-            arrowModeOptions.forEach(option => {
-                option.addEventListener('click', () => {
-                    const mode = option.getAttribute('data-mode');
-                    whiteboard.setArrowMode(mode);
-
-                    // Update tooltip to show current mode
-                    const modeIcons = {
-                        'end': '→',
-                        'start': '←',
-                        'both': '↔',
-                        'none': '—'
-                    };
-                    arrowBtn.setAttribute('data-tooltip', `Arrow (${modeIcons[mode]})`);
-
-                    arrowModeMenu.style.display = 'none';
-                });
-            });
-        }
-
-        // Math tools
-        const gridBtn = document.getElementById('tool-grid');
-        if (gridBtn) {
-            gridBtn.addEventListener('click', () => {
-                whiteboard.addCoordinateGrid();
-            });
-        }
-
-        const graphBtn = document.getElementById('tool-graph');
-        if (graphBtn) {
-            graphBtn.addEventListener('click', () => {
-                const funcStr = prompt('Enter function (e.g., x^2, 2x+1, Math.sin(x)):');
-                if (funcStr) {
-                    // Automatically add coordinate grid if not present
-                    whiteboard.addCoordinateGrid();
-                    // Then plot the function
-                    whiteboard.plotFunction(funcStr);
-                }
-            });
-        }
-
-        const protractorBtn = document.getElementById('tool-protractor');
-        if (protractorBtn) {
-            protractorBtn.addEventListener('click', () => {
-                whiteboard.addProtractor(whiteboard.canvas.width / 2, whiteboard.canvas.height / 2);
-            });
-        }
-
-        // Background menu dropdown
-        const backgroundMenuBtn = document.getElementById('background-menu-btn');
-        const backgroundMenu = document.getElementById('background-menu');
-        const bgUploadInput = document.getElementById('whiteboard-bg-upload');
-
-        if (backgroundMenuBtn && backgroundMenu) {
-            // Toggle dropdown
-            backgroundMenuBtn.addEventListener('click', (e) => {
-                e.stopPropagation();
-                backgroundMenu.style.display = backgroundMenu.style.display === 'none' ? 'block' : 'none';
-            });
-
-            // Close dropdown when clicking outside
-            document.addEventListener('click', (e) => {
-                if (!backgroundMenuBtn.contains(e.target) && !backgroundMenu.contains(e.target)) {
-                    backgroundMenu.style.display = 'none';
-                }
-            });
-
-            // Handle background option clicks
-            const bgOptions = backgroundMenu.querySelectorAll('.bg-option');
-            bgOptions.forEach(option => {
-                option.addEventListener('click', () => {
-                    const bgType = option.getAttribute('data-bg');
-
-                    if (bgType === 'upload') {
-                        bgUploadInput.click();
-                    } else {
-                        whiteboard.setPresetBackground(bgType);
-                    }
-
-                    backgroundMenu.style.display = 'none';
-                });
-            });
-
-            // File upload handler
-            if (bgUploadInput) {
-                bgUploadInput.addEventListener('change', (e) => {
-                    const file = e.target.files[0];
-                    if (file) {
-                        if (file.type.startsWith('image/')) {
-                            whiteboard.setBackgroundImage(file);
-                        } else if (file.type === 'application/pdf') {
-                            alert('PDF support coming soon! For now, please convert to an image or take a screenshot.');
-                        } else {
-                            alert('Please upload an image file (PNG, JPG, etc.)');
-                        }
-                    }
-                    // Reset input
-                    bgUploadInput.value = '';
-                });
-            }
-        }
-
-        // Color picker
-        const colorPicker = document.getElementById('color-picker');
-        const colorDisplay = document.querySelector('.color-display');
-        if (colorPicker && colorDisplay) {
-            colorPicker.addEventListener('change', (e) => {
-                whiteboard.setColor(e.target.value);
-                colorDisplay.style.color = e.target.value;
-            });
-        }
-
-        // Stroke width
-        const strokeSlider = document.getElementById('stroke-width-slider');
-        if (strokeSlider) {
-            strokeSlider.addEventListener('input', (e) => {
-                whiteboard.setStrokeWidth(e.target.value);
-            });
-        }
-
-        // Action buttons
-        const undoBtn = document.getElementById('undo-btn');
-        if (undoBtn) {
-            undoBtn.addEventListener('click', () => whiteboard.undo());
-        }
-
-        const redoBtn = document.getElementById('redo-btn');
-        if (redoBtn) {
-            redoBtn.addEventListener('click', () => whiteboard.redo());
-        }
-
-        const deleteBtn = document.getElementById('delete-btn');
-        if (deleteBtn) {
-            deleteBtn.addEventListener('click', () => whiteboard.deleteSelected());
-        }
-
-        const clearBtn = document.getElementById('clear-btn');
-        if (clearBtn) {
-            clearBtn.addEventListener('click', () => {
-                if (confirm('Clear entire whiteboard?')) {
-                    whiteboard.clear();
-                }
-            });
-        }
-
-        // Export menu with dropdown
-        const downloadBtn = document.getElementById('download-btn');
-        const exportMenu = document.getElementById('export-menu');
-        if (downloadBtn && exportMenu) {
-            // Toggle menu on click
-            downloadBtn.addEventListener('click', (e) => {
-                e.stopPropagation();
-                exportMenu.style.display = exportMenu.style.display === 'none' ? 'block' : 'none';
-            });
-
-            // Handle export options
-            document.querySelectorAll('.export-option').forEach(option => {
-                option.addEventListener('click', (e) => {
-                    const exportType = e.currentTarget.dataset.export;
-                    switch(exportType) {
-                        case 'png':
-                            whiteboard.exportToPNG();
-                            break;
-                        case 'pdf':
-                            whiteboard.exportToPDF();
-                            break;
-                        case 'clipboard':
-                            whiteboard.copyToClipboard();
-                            break;
-                    }
-                    exportMenu.style.display = 'none';
-                });
-            });
-
-            // Close menu when clicking outside
-            document.addEventListener('click', (e) => {
-                if (!downloadBtn.contains(e.target) && !exportMenu.contains(e.target)) {
-                    exportMenu.style.display = 'none';
-                }
-            });
-        }
-
-        // Region overlay toggle
-        const regionOverlayBtn = document.getElementById('region-overlay-btn');
-        if (regionOverlayBtn) {
-            regionOverlayBtn.addEventListener('click', () => {
-                whiteboard.toggleRegionOverlay();
-                regionOverlayBtn.classList.toggle('active');
-            });
-        }
-
-        // Shortcuts help
-        const shortcutsHelpBtn = document.getElementById('shortcuts-help-btn');
-        if (shortcutsHelpBtn) {
-            shortcutsHelpBtn.addEventListener('click', () => {
-                whiteboard.toggleShortcutsPanel();
-            });
-        }
-
-        const sendToAiBtn = document.getElementById('send-to-ai-btn');
-        if (sendToAiBtn) {
-            sendToAiBtn.addEventListener('click', async () => {
-                try {
-                    // Convert Fabric.js canvas to data URL
-                    const dataURL = whiteboard.canvas.toDataURL({
-                        format: 'png',
-                        quality: 1,
-                        multiplier: 2 // Higher resolution for better OCR
-                    });
-
-                    // Convert data URL to blob
-                    const response = await fetch(dataURL);
-                    const blob = await response.blob();
-
-                    if (!blob) {
-                        alert('Failed to capture whiteboard. Please try again.');
-                        return;
-                    }
-
-                    // Create file from blob
-                    const file = new File([blob], 'whiteboard-drawing.png', { type: 'image/png' });
-
-                    // Use the proper file upload handler to add to attachedFiles array
-                    if (window.handleFileUpload) {
-                        window.handleFileUpload(file);
-                    } else {
-                        console.error('handleFileUpload not found - falling back to direct attachment');
-                        attachedFile = file; // Fallback for backwards compatibility
-                        showFilePill(file.name);
-                    }
-
-                    // Set placeholder message (user can edit or replace)
-                    if (!userInput.textContent.trim()) {
-                        userInput.textContent = 'Can you help me with this?';
-                    }
-
-                    // Focus input so user can type or send
-                    userInput.focus();
-                    userInput.select();
-
-                    console.log('✅ Whiteboard screenshot attached! Click send when ready.');
-                } catch (error) {
-                    console.error('Error capturing whiteboard:', error);
-                    alert('Failed to capture whiteboard. Please try again.');
-                }
-            });
-        }
-
-        // Panel controls are now handled in whiteboard.js setupPanelControls()
-        // Open whiteboard button (when closed)
-        const openWhiteboardBtn = document.getElementById('open-whiteboard-btn');
-        if (openWhiteboardBtn) {
-            openWhiteboardBtn.addEventListener('click', () => {
-                whiteboard.show();
-                openWhiteboardBtn.classList.add('hidden');
-            });
-        }
-
-        // Keyboard shortcuts now handled in whiteboard.js setupKeyboardShortcuts()
-    }
-
-    function updateDrawingMode() {
-        if (whiteboardState.currentTool === 'pen') {
-            fabricCanvas.isDrawingMode = true;
-            fabricCanvas.selection = false;
-            fabricCanvas.freeDrawingBrush.color = whiteboardState.currentColor; // Restore current color
-            fabricCanvas.freeDrawingBrush.width = whiteboardState.brushSize;
-        } else if (whiteboardState.currentTool === 'eraser') {
-            fabricCanvas.isDrawingMode = true;
-            fabricCanvas.freeDrawingBrush.color = '#ffffff'; // White eraser
-            fabricCanvas.freeDrawingBrush.width = whiteboardState.brushSize * 2; // Larger eraser
-            fabricCanvas.selection = false;
-        } else {
-            // Shape tools (line, circle, rectangle, text)
-            fabricCanvas.isDrawingMode = false;
-            fabricCanvas.selection = false;
-        }
-    }
-
-    function setupWhiteboardDrawing() {
-        fabricCanvas.on('mouse:down', function(options) {
-            if (fabricCanvas.isDrawingMode) return; // Let free drawing handle it
-
-            const pointer = fabricCanvas.getPointer(options.e);
-            whiteboardState.isDrawing = true;
-            whiteboardState.startX = pointer.x;
-            whiteboardState.startY = pointer.y;
-
-            switch (whiteboardState.currentTool) {
-                case 'line':
-                    whiteboardState.currentShape = new fabric.Line(
-                        [pointer.x, pointer.y, pointer.x, pointer.y],
-                        {
-                            stroke: whiteboardState.currentColor,
-                            strokeWidth: whiteboardState.brushSize,
-                            selectable: false
-                        }
-                    );
-                    fabricCanvas.add(whiteboardState.currentShape);
-                    break;
-
-                case 'circle':
-                    whiteboardState.currentShape = new fabric.Circle({
-                        left: pointer.x,
-                        top: pointer.y,
-                        radius: 1,
-                        stroke: whiteboardState.currentColor,
-                        strokeWidth: whiteboardState.brushSize,
-                        fill: 'transparent',
-                        selectable: false,
-                        originX: 'center',
-                        originY: 'center'
-                    });
-                    fabricCanvas.add(whiteboardState.currentShape);
-                    break;
-
-                case 'rectangle':
-                    whiteboardState.currentShape = new fabric.Rect({
-                        left: pointer.x,
-                        top: pointer.y,
-                        width: 1,
-                        height: 1,
-                        stroke: whiteboardState.currentColor,
-                        strokeWidth: whiteboardState.brushSize,
-                        fill: 'transparent',
-                        selectable: false
-                    });
-                    fabricCanvas.add(whiteboardState.currentShape);
-                    break;
-
-                case 'text':
-                    const text = prompt('Enter text:');
-                    if (text) {
-                        const textObj = new fabric.Text(text, {
-                            left: pointer.x,
-                            top: pointer.y,
-                            fill: whiteboardState.currentColor,
-                            fontSize: whiteboardState.brushSize * 5,
-                            selectable: false
-                        });
-                        fabricCanvas.add(textObj);
-                    }
-                    whiteboardState.isDrawing = false;
-                    break;
-            }
-        });
-
-        fabricCanvas.on('mouse:move', function(options) {
-            if (!whiteboardState.isDrawing || !whiteboardState.currentShape) return;
-
-            const pointer = fabricCanvas.getPointer(options.e);
-
-            switch (whiteboardState.currentTool) {
-                case 'line':
-                    whiteboardState.currentShape.set({
-                        x2: pointer.x,
-                        y2: pointer.y
-                    });
-                    break;
-
-                case 'circle':
-                    const radius = Math.sqrt(
-                        Math.pow(pointer.x - whiteboardState.startX, 2) +
-                        Math.pow(pointer.y - whiteboardState.startY, 2)
-                    );
-                    whiteboardState.currentShape.set({ radius: radius });
-                    break;
-
-                case 'rectangle':
-                    whiteboardState.currentShape.set({
-                        width: Math.abs(pointer.x - whiteboardState.startX),
-                        height: Math.abs(pointer.y - whiteboardState.startY)
-                    });
-                    if (pointer.x < whiteboardState.startX) {
-                        whiteboardState.currentShape.set({ left: pointer.x });
-                    }
-                    if (pointer.y < whiteboardState.startY) {
-                        whiteboardState.currentShape.set({ top: pointer.y });
-                    }
-                    break;
-            }
-
-            fabricCanvas.renderAll();
-        });
-
-        fabricCanvas.on('mouse:up', function() {
-            whiteboardState.isDrawing = false;
-            whiteboardState.currentShape = null;
-        });
-    }
-
-    async function shareWhiteboardWithAI() {
-        if (!fabricCanvas || !currentUser) return;
-
-        try {
-            // Convert canvas to base64 image
-            const imageData = fabricCanvas.toDataURL({
-                format: 'png',
-                quality: 0.8
-            });
-
-            // Convert base64 to blob
-            const response = await fetch(imageData);
-            const blob = await response.blob();
-            const file = new File([blob], 'whiteboard.png', { type: 'image/png' });
-
-            // Use the existing file upload mechanism
-            attachedFile = file;
-
-            // Create visual pill
-            const filePillContainer = document.getElementById('file-pill-container');
-            filePillContainer.innerHTML = `
-                <div class="file-pill">
-                    <span class="file-name">📋 Whiteboard snapshot</span>
-                    <button class="remove-file-btn" onclick="removeAttachedFile()">×</button>
-                </div>
-            `;
-
-            // Suggest a message
-            const userInput = document.getElementById('user-input');
-            if (userInput && !userInput.textContent.trim()) {
-                userInput.textContent = "Can you help me with this problem I drew on the whiteboard?";
-            }
-
-            showToast('Whiteboard snapshot attached! Click send to share with AI.', 3000);
-        } catch (error) {
-            console.error('Error sharing whiteboard:', error);
-            showToast('Failed to share whiteboard. Please try again.', 3000);
-        }
-    }
-
-    function drawBackgroundTemplate(templateType) {
-        if (!fabricCanvas) return;
-
-        // Clear only background objects (preserve user drawings)
-        const objects = fabricCanvas.getObjects();
-        objects.forEach(obj => {
-            if (obj.isBackground) {
-                fabricCanvas.remove(obj);
-            }
-        });
-
-        if (templateType === 'none') {
-            fabricCanvas.renderAll();
-            return;
-        }
-
-        const width = fabricCanvas.width;
-        const height = fabricCanvas.height;
-        const gridColor = '#e0e0e0';
-        const axisColor = '#12B3B3';
-        const labelColor = '#666';
-
-        switch (templateType) {
-            case 'coordinate-grid':
-                drawCoordinateGrid(width, height, gridColor, axisColor, labelColor);
-                break;
-            case 'number-line':
-                drawNumberLine(width, height, axisColor, labelColor);
-                break;
-            case 'graph-paper':
-                drawGraphPaper(width, height, gridColor);
-                break;
-            case 'algebra-tiles':
-                drawAlgebraTilesMat(width, height, gridColor);
-                break;
-        }
-
-        fabricCanvas.renderAll();
-    }
-
-    function drawCoordinateGrid(width, height, gridColor, axisColor, labelColor) {
-        const centerX = width / 2;
-        const centerY = height / 2;
-        const spacing = Math.min(width, height) / 20; // 20 units total (-10 to 10)
-
-        // Draw grid lines
-        for (let i = -10; i <= 10; i++) {
-            // Vertical lines
-            const x = centerX + (i * spacing);
-            const vLine = new fabric.Line([x, 0, x, height], {
-                stroke: gridColor,
-                strokeWidth: 1,
-                selectable: false,
-                evented: false,
-                isBackground: true
-            });
-            fabricCanvas.add(vLine);
-
-            // Horizontal lines
-            const y = centerY + (i * spacing);
-            const hLine = new fabric.Line([0, y, width, y], {
-                stroke: gridColor,
-                strokeWidth: 1,
-                selectable: false,
-                evented: false,
-                isBackground: true
-            });
-            fabricCanvas.add(hLine);
-        }
-
-        // Draw axes
-        const xAxis = new fabric.Line([0, centerY, width, centerY], {
-            stroke: axisColor,
-            strokeWidth: 2,
-            selectable: false,
-            evented: false,
-            isBackground: true
-        });
-        fabricCanvas.add(xAxis);
-
-        const yAxis = new fabric.Line([centerX, 0, centerX, height], {
-            stroke: axisColor,
-            strokeWidth: 2,
-            selectable: false,
-            evented: false,
-            isBackground: true
-        });
-        fabricCanvas.add(yAxis);
-
-        // Draw axis labels
-        for (let i = -10; i <= 10; i += 2) {
-            if (i === 0) continue;
-
-            // X-axis labels
-            const xLabelPos = centerX + (i * spacing);
-            const xLabel = new fabric.Text(i.toString(), {
-                left: xLabelPos,
-                top: centerY + 5,
-                fontSize: 12,
-                fill: labelColor,
-                selectable: false,
-                evented: false,
-                isBackground: true,
-                originX: 'center'
-            });
-            fabricCanvas.add(xLabel);
-
-            // Y-axis labels
-            const yLabelPos = centerY - (i * spacing);
-            const yLabel = new fabric.Text(i.toString(), {
-                left: centerX - 20,
-                top: yLabelPos,
-                fontSize: 12,
-                fill: labelColor,
-                selectable: false,
-                evented: false,
-                isBackground: true,
-                originY: 'center'
-            });
-            fabricCanvas.add(yLabel);
-        }
-
-        // Origin label
-        const origin = new fabric.Text('0', {
-            left: centerX - 15,
-            top: centerY + 5,
-            fontSize: 12,
-            fill: labelColor,
-            selectable: false,
-            evented: false,
-            isBackground: true
-        });
-        fabricCanvas.add(origin);
-    }
-
-    function drawNumberLine(width, height, axisColor, labelColor) {
-        const centerY = height / 2;
-        const spacing = width / 22; // Padding on sides
-        const startX = spacing;
-        const endX = width - spacing;
-        const unit = (endX - startX) / 20;
-
-        // Draw main line
-        const line = new fabric.Line([startX, centerY, endX, centerY], {
-            stroke: axisColor,
-            strokeWidth: 3,
-            selectable: false,
-            evented: false,
-            isBackground: true
-        });
-        fabricCanvas.add(line);
-
-        // Draw tick marks and labels
-        for (let i = -10; i <= 10; i++) {
-            const x = startX + (i + 10) * unit;
-            const tickHeight = i % 5 === 0 ? 15 : 10;
-
-            const tick = new fabric.Line([x, centerY - tickHeight, x, centerY + tickHeight], {
-                stroke: axisColor,
-                strokeWidth: 2,
-                selectable: false,
-                evented: false,
-                isBackground: true
-            });
-            fabricCanvas.add(tick);
-
-            if (i % 2 === 0) {
-                const label = new fabric.Text(i.toString(), {
-                    left: x,
-                    top: centerY + 25,
-                    fontSize: 14,
-                    fill: labelColor,
-                    selectable: false,
-                    evented: false,
-                    isBackground: true,
-                    originX: 'center'
-                });
-                fabricCanvas.add(label);
-            }
-        }
-
-        // Add arrows
-        const arrowSize = 10;
-        const leftArrow = new fabric.Triangle({
-            left: startX - arrowSize,
-            top: centerY,
-            width: arrowSize,
-            height: arrowSize,
-            fill: axisColor,
-            angle: -90,
-            selectable: false,
-            evented: false,
-            isBackground: true,
-            originX: 'center',
-            originY: 'center'
-        });
-        fabricCanvas.add(leftArrow);
-
-        const rightArrow = new fabric.Triangle({
-            left: endX + arrowSize,
-            top: centerY,
-            width: arrowSize,
-            height: arrowSize,
-            fill: axisColor,
-            angle: 90,
-            selectable: false,
-            evented: false,
-            isBackground: true,
-            originX: 'center',
-            originY: 'center'
-        });
-        fabricCanvas.add(rightArrow);
-    }
-
-    function drawGraphPaper(width, height, gridColor) {
-        const spacing = 20; // 20 pixels per square
-
-        // Draw vertical lines
-        for (let x = 0; x <= width; x += spacing) {
-            const line = new fabric.Line([x, 0, x, height], {
-                stroke: gridColor,
-                strokeWidth: 1,
-                selectable: false,
-                evented: false,
-                isBackground: true
-            });
-            fabricCanvas.add(line);
-        }
-
-        // Draw horizontal lines
-        for (let y = 0; y <= height; y += spacing) {
-            const line = new fabric.Line([0, y, width, y], {
-                stroke: gridColor,
-                strokeWidth: 1,
-                selectable: false,
-                evented: false,
-                isBackground: true
-            });
-            fabricCanvas.add(line);
-        }
-    }
-
-    function drawAlgebraTilesMat(width, height, gridColor) {
-        const padding = 40;
-        const matWidth = width - (2 * padding);
-        const matHeight = height - (2 * padding);
-
-        // Draw mat border
-        const border = new fabric.Rect({
-            left: padding,
-            top: padding,
-            width: matWidth,
-            height: matHeight,
-            stroke: '#12B3B3',
-            strokeWidth: 3,
-            fill: 'transparent',
-            selectable: false,
-            evented: false,
-            isBackground: true
-        });
-        fabricCanvas.add(border);
-
-        // Draw positive/negative divider (horizontal)
-        const divider = new fabric.Line([padding, height / 2, width - padding, height / 2], {
-            stroke: '#12B3B3',
-            strokeWidth: 2,
-            strokeDashArray: [5, 5],
-            selectable: false,
-            evented: false,
-            isBackground: true
-        });
-        fabricCanvas.add(divider);
-
-        // Add labels
-        const posLabel = new fabric.Text('Positive', {
-            left: width / 2,
-            top: padding + 15,
-            fontSize: 16,
-            fill: '#16C86D',
-            fontWeight: 'bold',
-            selectable: false,
-            evented: false,
-            isBackground: true,
-            originX: 'center'
-        });
-        fabricCanvas.add(posLabel);
-
-        const negLabel = new fabric.Text('Negative', {
-            left: width / 2,
-            top: height / 2 + 15,
-            fontSize: 16,
-            fill: '#FF4E4E',
-            fontWeight: 'bold',
-            selectable: false,
-            evented: false,
-            isBackground: true,
-            originX: 'center'
-        });
-        fabricCanvas.add(negLabel);
-
-        // Draw light grid
-        const gridSpacing = 30;
-        for (let x = padding; x <= width - padding; x += gridSpacing) {
-            const line = new fabric.Line([x, padding, x, height - padding], {
-                stroke: '#f0f0f0',
-                strokeWidth: 1,
-                selectable: false,
-                evented: false,
-                isBackground: true
-            });
-            fabricCanvas.add(line);
-        }
-        for (let y = padding; y <= height - padding; y += gridSpacing) {
-            const line = new fabric.Line([padding, y, width - padding, y], {
-                stroke: '#f0f0f0',
-                strokeWidth: 1,
-                selectable: false,
-                evented: false,
-                isBackground: true
-            });
-            fabricCanvas.add(line);
-        }
-    }
-
-    // Transform mathematical coordinates to canvas pixel coordinates
-    function mathToCanvasCoords(mathX, mathY, mathMin = -10, mathMax = 10) {
-        if (!fabricCanvas) return { x: 0, y: 0 };
-
-        const canvasWidth = fabricCanvas.width;
-        const canvasHeight = fabricCanvas.height;
-        const padding = 40; // Padding from canvas edges
-
-        // Available drawing area
-        const drawWidth = canvasWidth - (2 * padding);
-        const drawHeight = canvasHeight - (2 * padding);
-
-        // Transform from math coordinates to pixel coordinates
-        const mathRange = mathMax - mathMin;
-        const pixelX = padding + ((mathX - mathMin) / mathRange) * drawWidth;
-        const pixelY = padding + ((mathMax - mathY) / mathRange) * drawHeight; // Flip Y axis
-
-        return { x: pixelX, y: pixelY };
-    }
-
-    async function renderDrawing(sequence, delay = 500) {
-        if (!whiteboard) return;
-        await whiteboard.renderAIDrawing(sequence, delay);
-    }
-
-    function makeElementDraggable(elmnt) {
-        let pos1 = 0, pos2 = 0, pos3 = 0, pos4 = 0;
-        const header = elmnt.querySelector(".dashboard-panel-header");
-
-        if (header) {
-            header.onmousedown = dragMouseDown;
-        } else {
-            elmnt.onmousedown = dragMouseDown;
-        }
-
-        function dragMouseDown(e) {
-            e = e || window.event;
-            e.preventDefault();
-            pos3 = e.clientX;
-            pos4 = e.clientY;
-            document.onmouseup = closeDragElement;
-            document.onmousemove = elementDrag;
-        }
-
-        function elementDrag(e) {
-            e = e || window.event;
-            e.preventDefault();
-            pos1 = pos3 - e.clientX;
-            pos2 = pos4 - e.clientY;
-            pos3 = e.clientX;
-            pos4 = e.clientY;
-            elmnt.style.top = (elmnt.offsetTop - pos2) + "px";
-            elmnt.style.left = (elmnt.offsetLeft - pos1) + "px";
-        }
-
-        function closeDragElement() {
-            document.onmouseup = null;
-            document.onmousemove = null;
-        }
-    }
 
   // ============================================
     function setupChatUI() {
@@ -3095,323 +1822,14 @@ document.addEventListener("DOMContentLoaded", () => {
         }
     }
 
-    /**
-     * Show assessment pitch with explanation and skip option
-     */
-    function showAssessmentPitch() {
-        const pitchMessage = `**Quick Placement Assessment** 📊
+    // --- Assessment System (see modules/assessment.js) ---
+    const assessment = createAssessmentSystem({
+        appendMessage,
+        showThinkingIndicator,
+        getChatBox: () => chatBox,
+    });
+    const { showAssessmentPitch } = assessment;
 
-This helps me personalize your learning by understanding:
-• What you already know well
-• Where you might need support
-• The perfect starting point for you
-
-**What to expect:**
-• 5-30 adaptive questions (stops when I'm confident)
-• Questions adjust to your level as you go
-• Takes about 5-10 minutes
-
-Ready to start? Or would you rather skip for now and jump straight into general tutoring?`;
-
-        appendMessage(pitchMessage, "ai");
-
-        // Add action buttons
-        const buttonContainer = document.createElement('div');
-        buttonContainer.className = 'assessment-action-buttons';
-        buttonContainer.style.cssText = `
-            display: flex;
-            gap: 12px;
-            margin: 15px 0;
-            flex-wrap: wrap;
-        `;
-
-        const startBtn = document.createElement('button');
-        startBtn.textContent = "Let's do it!";
-        startBtn.className = 'btn-primary';
-        startBtn.style.cssText = `
-            padding: 12px 24px;
-            background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-            color: white;
-            border: none;
-            border-radius: 8px;
-            font-weight: 600;
-            cursor: pointer;
-            transition: transform 0.2s;
-        `;
-        startBtn.onmouseover = () => startBtn.style.transform = 'scale(1.05)';
-        startBtn.onmouseout = () => startBtn.style.transform = 'scale(1)';
-        startBtn.onclick = () => {
-            buttonContainer.remove();
-            startInChatAssessment();
-        };
-
-        const skipBtn = document.createElement('button');
-        skipBtn.textContent = "Skip for now";
-        skipBtn.className = 'btn-secondary';
-        skipBtn.style.cssText = `
-            padding: 12px 24px;
-            background: #f3f4f6;
-            color: #6b7280;
-            border: 1px solid #d1d5db;
-            border-radius: 8px;
-            font-weight: 600;
-            cursor: pointer;
-            transition: all 0.2s;
-        `;
-        skipBtn.onmouseover = () => {
-            skipBtn.style.background = '#e5e7eb';
-            skipBtn.style.color = '#4b5563';
-        };
-        skipBtn.onmouseout = () => {
-            skipBtn.style.background = '#f3f4f6';
-            skipBtn.style.color = '#6b7280';
-        };
-        skipBtn.onclick = async () => {
-            buttonContainer.remove();
-            appendMessage("No problem! I'll ask again next time. What would you like to work on today?", "ai");
-            // Mark that user skipped (will be offered again next login)
-            try {
-                await csrfFetch('/api/screener/skip', {
-                    method: 'POST',
-                    headers: { 'Content-Type': 'application/json' }
-                });
-            } catch (error) {
-                console.error('Failed to record skip:', error);
-            }
-        };
-
-        buttonContainer.appendChild(startBtn);
-        buttonContainer.appendChild(skipBtn);
-        chatBox.appendChild(buttonContainer);
-        chatBox.scrollTop = chatBox.scrollHeight;
-    }
-
-    /**
-     * Start in-chat CAT assessment
-     */
-    async function startInChatAssessment() {
-        try {
-            appendMessage("Great! Starting your placement assessment...", "ai");
-            showThinkingIndicator(true);
-
-            // Start screener session
-            const response = await csrfFetch('/api/screener/start', {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' }
-            });
-
-            if (!response.ok) {
-                // Get error details from response
-                const errorData = await response.json();
-                const errorMsg = errorData.message || errorData.error || 'Failed to start assessment';
-                throw new Error(errorMsg);
-            }
-
-            const data = await response.json();
-            console.log('[Assessment] Started:', data);
-
-            // Load first problem
-            await loadNextAssessmentProblem(data.sessionId);
-
-        } catch (error) {
-            console.error('[Assessment] Start error:', error);
-            appendMessage(`Oops, something went wrong starting the assessment: ${error.message}. Let's just chat instead - what would you like to work on?`, "ai");
-            showThinkingIndicator(false);
-        }
-    }
-
-    /**
-     * Load next assessment problem
-     */
-    async function loadNextAssessmentProblem(sessionId) {
-        try {
-            const response = await csrfFetch(`/api/screener/next-problem?sessionId=${sessionId}`, {
-                credentials: 'include'
-            });
-
-            if (!response.ok) {
-                // Get error details from response
-                const errorData = await response.json();
-                const errorMsg = errorData.message || errorData.error || 'Failed to load problem';
-                throw new Error(errorMsg);
-            }
-
-            const data = await response.json();
-            showThinkingIndicator(false);
-
-            if (data.complete) {
-                // Assessment finished - pass sessionId
-                data.sessionId = sessionId;
-                handleAssessmentComplete(data);
-                return;
-            }
-
-            // Display problem
-            const problemMessage = `**Question ${data.problem.questionNumber} of ~${data.problem.progress?.target || '?'}**
-
-${data.problem.content}`;
-
-            appendMessage(problemMessage, "ai");
-
-            // Create answer input
-            createAssessmentAnswerInput(sessionId, data.problem);
-
-        } catch (error) {
-            console.error('[Assessment] Load problem error:', error);
-            appendMessage(`Had trouble loading the next question: ${error.message}. Let's continue with regular tutoring instead!`, "ai");
-            showThinkingIndicator(false);
-        }
-    }
-
-    /**
-     * Create answer input for assessment
-     */
-    function createAssessmentAnswerInput(sessionId, problem) {
-        const inputContainer = document.createElement('div');
-        inputContainer.className = 'assessment-answer-input';
-        inputContainer.style.cssText = `
-            display: flex;
-            gap: 12px;
-            margin: 15px 0;
-            align-items: center;
-        `;
-
-        const input = document.createElement('input');
-        input.type = 'text';
-        input.placeholder = 'Your answer...';
-        input.style.cssText = `
-            flex: 1;
-            padding: 12px;
-            border: 2px solid #e5e7eb;
-            border-radius: 8px;
-            font-size: 16px;
-        `;
-
-        const submitBtn = document.createElement('button');
-        submitBtn.textContent = 'Submit';
-        submitBtn.style.cssText = `
-            padding: 12px 24px;
-            background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-            color: white;
-            border: none;
-            border-radius: 8px;
-            font-weight: 600;
-            cursor: pointer;
-        `;
-
-        submitBtn.onclick = async () => {
-            const answer = input.value.trim();
-            if (!answer) return;
-
-            inputContainer.remove();
-            await submitAssessmentAnswer(sessionId, problem.problemId, answer);
-        };
-
-        input.onkeypress = (e) => {
-            if (e.key === 'Enter') submitBtn.click();
-        };
-
-        inputContainer.appendChild(input);
-        inputContainer.appendChild(submitBtn);
-        chatBox.appendChild(inputContainer);
-        chatBox.scrollTop = chatBox.scrollHeight;
-        input.focus();
-    }
-
-    /**
-     * Submit assessment answer
-     */
-    async function submitAssessmentAnswer(sessionId, problemId, answer) {
-        try {
-            appendMessage(answer, "user");
-            showThinkingIndicator(true);
-
-            const response = await csrfFetch('/api/screener/submit-answer', {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({
-                    sessionId,
-                    problemId,
-                    answer,
-                    responseTime: 0 // Could track this if needed
-                })
-            });
-
-            if (!response.ok) {
-                // Get error details from response
-                const errorData = await response.json();
-                const errorMsg = errorData.message || errorData.error || 'Failed to submit answer';
-                throw new Error(errorMsg);
-            }
-
-            const data = await response.json();
-
-            // Show feedback
-            if (data.feedback) {
-                appendMessage(data.feedback, "ai");
-            }
-
-            // Load next problem or finish
-            if (data.nextAction === 'interview' || data.nextAction === 'complete') {
-                // Assessment reached stopping criteria - complete it
-                data.sessionId = sessionId;
-                handleAssessmentComplete(data);
-            } else {
-                // Continue with next problem
-                await loadNextAssessmentProblem(sessionId);
-            }
-
-        } catch (error) {
-            console.error('[Assessment] Submit error:', error);
-            showThinkingIndicator(false);
-            appendMessage(`Had trouble processing that: ${error.message}. Let's move on to regular tutoring!`, "ai");
-        }
-    }
-
-    /**
-     * Handle assessment completion
-     */
-    async function handleAssessmentComplete(data) {
-        showThinkingIndicator(false);
-
-        // Call the complete endpoint to save results
-        try {
-            const completeResponse = await csrfFetch('/api/screener/complete', {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({
-                    sessionId: data.sessionId
-                })
-            });
-
-            if (completeResponse.ok) {
-                const completeData = await completeResponse.json();
-                const report = completeData.report;
-
-                const badgeInfo = report.earnedBadges && report.earnedBadges.length > 0
-                    ? `\n• **Badges Earned**: ${report.earnedBadges.length} 🎖️`
-                    : '';
-
-                const resultsMessage = `🎉 **Assessment Complete!**
-
-You did great! Here's what I learned:
-• **Level**: θ = ${report.theta.toFixed(2)} (${report.percentile}th percentile)
-• **Accuracy**: ${Math.round(report.accuracy * 100)}%
-• **Questions**: ${report.questionsAnswered}${badgeInfo}
-
-I now have a good understanding of where you are. I'll use this to personalize everything for you!
-
-What would you like to work on first?`;
-
-                appendMessage(resultsMessage, "ai");
-            } else {
-                throw new Error('Failed to complete assessment');
-            }
-        } catch (error) {
-            console.error('Failed to save assessment results:', error);
-            appendMessage("Assessment complete! What would you like to work on?", "ai");
-        }
-    }
 
     // Audio, gamification display, leaderboard, and quests are now in ES modules
     // (see imports at top of file)
